Init update to 2.0.0 -> Breaks old plugin... and table with balance.

+++ src/main/kotlin/com/github/encryptsl/lite/eco/common/database/entity/User.kt
@@ -1,5 +1,6 @@
 package com.github.encryptsl.lite.eco.common.database.entity
 
+import java.math.BigDecimal
 import java.util.UUID
 
-data class User(val userName: String, val uuid: UUID, val money: Double)
+data class User(val userName: String, val uuid: UUID, val money: BigDecimal)
